+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.0.3
+
+* Adds Hebrew to built-in tables (now includes English and Hebrew)
+* Adds support for custom tables
+* Requires specifying a table when instantiating
+* Changes api to use Gematria::Calculator (this may still not be stable)
+
 ## v0.0.2
 
 * Ruby 1.8.7 compatible (Now compatible with 1.8.7 - head)